webhook_proxy/requirements.txt
fastapi~=0.115.11
requests
uvicorn
python-dotenv~=1.0.1